+++ app/src/main/java/eu/kanade/tachiyomi/ui/migration/MigrationController.kt
@@ -15,7 +15,7 @@ import eu.kanade.tachiyomi.ui.base.controller.withFadeTransaction
 import eu.kanade.tachiyomi.ui.migration.manga.design.PreMigrationController
 import eu.kanade.tachiyomi.ui.migration.manga.process.MigrationListController
 import eu.kanade.tachiyomi.util.view.RecyclerWindowInsetsListener
-import eu.kanade.tachiyomi.util.await
+import eu.kanade.tachiyomi.util.system.await
 import eu.kanade.tachiyomi.util.system.launchUI
 import eu.kanade.tachiyomi.ui.migration.manga.process.MigrationProcedureConfig
 import kotlinx.android.synthetic.main.migration_controller.*
