+++ app/src/main/java/eu/kanade/tachiyomi/util/system/RxUtil.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.util
+package eu.kanade.tachiyomi.util.system
 
 import kotlinx.coroutines.suspendCancellableCoroutine
 import rx.Scheduler
